[#3232 #3236] Fix patchMultiple()
patchMultiple() needs to run modifyPeriods(false) as a data prep callback for "regular" results frameworks, so that we copy period.locked to period.is_locked for all changed periods.
